moving router to its own pkg; running go mod init
