+++ dask-gateway-server/dask-gateway-proxy/go.mod
@@ -1,3 +1,11 @@
 module github.com/dask/dask-gateway/dask-gateway-proxy
 
 go 1.18
+
+require github.com/stretchr/testify v1.7.1
+
+require (
+	github.com/davecgh/go-spew v1.1.0 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	gopkg.in/yaml.v3 v3.0.0-20200313102051-9f266ea9e77c // indirect
+)
