+++ dask-gateway-server/dask-gateway-proxy/pkg/router/router.go
@@ -1,4 +1,4 @@
-package main
+package router
 
 import (
 	"encoding/json"
