+++ dask-gateway-server/dask-gateway-proxy/pkg/router/router_test.go
@@ -1,4 +1,4 @@
-package main
+package router
 
 import (
 	"fmt"
